syscalls: Add test for close-on-fork
Adds a test for new close-on-fork functionality. The first
iteration tests that the close-on-fork flag is not set and
that the child process can write to a temp file. The second
iteration tests that the close-on-fork flag is successfully
set using fcntl(2) and that the child process cannot write
to the temp file.

Contains a TODO to update the required kernel version once
the close-on-fork changes are included in a release.

Includes sys/wait.h in tst_safe_macros.h to get declaration
for waitpid(2).

Signed-off-by: Nate Karstens <nate.karstens@garmin.com>