syscalls: Test close-on-fork with unshare(2)
Adds a test for how close-on-fork behaves with unshare(2). A
child process is created using clone(2), with files set to be
shared between parent and child. If the close-on-fork flag is
not set, then both parent and child each receive their own copy
of the file descriptor and can both write to the file. If the
close-on-fork flag is set, then the process calling unshare(2)
will not receive a copy of the file descriptor and so will fail
to write to the file.

Contains a TODO to update the required kernel version once
the close-on-fork changes are included in a release.

Signed-off-by: Nate Karstens <nate.karstens@garmin.com>